% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{add_cases}
\alias{add_cases}
\title{Add test cases to a dataset}
\usage{
add_cases(dataset, ...)
}
\arguments{
\item{dataset}{A \code{secureeval_dataset} object.}

\item{...}{One or more \code{secureeval_test_case} objects to add.}
}
\value{
A new \code{secureeval_dataset} with the additional cases.
}
\description{
Add test cases to a dataset
}
