% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_size}
\alias{dataset_size}
\title{Get the number of cases in a dataset}
\usage{
dataset_size(dataset)
}
\arguments{
\item{dataset}{A \code{secureeval_dataset} object.}
}
\value{
An integer giving the number of test cases.
}
\description{
Get the number of cases in a dataset
}
