% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset_summary}
\alias{dataset_summary}
\title{Summarize a dataset by label}
\usage{
dataset_summary(dataset)
}
\arguments{
\item{dataset}{A \code{secureeval_dataset} object.}
}
\value{
A named list with counts per label.
}
\description{
Summarize a dataset by label
}
