% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{eval_contains}
\alias{eval_contains}
\title{Built-in evaluator: contains}
\usage{
eval_contains(case_sensitive = TRUE)
}
\arguments{
\item{case_sensitive}{Whether the comparison is case-sensitive. Defaults to \code{TRUE}.}
}
\value{
A \code{secureeval_evaluator} object.
}
\description{
Scores 1 if expected is found within result (both coerced to character).
}
\examples{
e <- eval_contains()
e$score_fn("hello world", "world") # 1
}
