% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{eval_custom}
\alias{eval_custom}
\title{Built-in evaluator: custom function}
\usage{
eval_custom(fn)
}
\arguments{
\item{fn}{A function \verb{function(result, expected)} returning a numeric value
between 0 and 1.}
}
\value{
A \code{secureeval_evaluator} object.
}
\description{
Wrap an arbitrary scoring function as an evaluator.
}
\examples{
e <- eval_custom(function(result, expected) {
  nchar(result) / nchar(expected)
})
}
