% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{eval_dataset}
\alias{eval_dataset}
\title{Create an evaluation dataset}
\usage{
eval_dataset(cases = list(), name = "", description = "")
}
\arguments{
\item{cases}{A list of \code{secureeval_test_case} objects.}

\item{name}{A name for the dataset.}

\item{description}{A description of the dataset.}
}
\value{
An object of class \code{secureeval_dataset}.
}
\description{
A dataset is a named collection of test cases.
}
\examples{
ds <- eval_dataset(
  cases = list(test_case("2+2", "4", label = "math")),
  name = "basic-math"
)
ds
}
