% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluators.R
\name{eval_regex_match}
\alias{eval_regex_match}
\title{Built-in evaluator: regex match}
\usage{
eval_regex_match()
}
\value{
A \code{secureeval_evaluator} object.
}
\description{
Scores 1 if result matches expected as a regular expression.
}
\examples{
e <- eval_regex_match()
e$score_fn("hello world", "^hello")
}
