% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{is_eval_run_result}
\alias{is_eval_run_result}
\title{Test if an object is an eval run result}
\usage{
is_eval_run_result(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{eval_run_result}, \code{FALSE} otherwise.
}
\description{
Test if an object is an eval run result
}
