% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\name{is_evaluator}
\alias{is_evaluator}
\title{Test if an object is an evaluator}
\usage{
is_evaluator(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{secureeval_evaluator}, \code{FALSE} otherwise.
}
\description{
Test if an object is an evaluator
}
