% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-case.R
\name{is_test_case}
\alias{is_test_case}
\title{Test if an object is a test case}
\usage{
is_test_case(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{secureeval_test_case}, \code{FALSE} otherwise.
}
\description{
Test if an object is a test case
}
