% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{load_dataset}
\alias{load_dataset}
\title{Load a dataset from JSON}
\usage{
load_dataset(path)
}
\arguments{
\item{path}{File path to load from.}
}
\value{
A \code{secureeval_dataset} object.
}
\description{
Load a dataset from JSON
}
