% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-reference.R
\name{reference_datasets}
\alias{reference_datasets}
\title{List the bundled datasets}
\usage{
reference_datasets()
}
\value{
A character vector of names you can pass to \code{\link[=load_reference]{load_reference()}}.
}
\description{
List the bundled datasets
}
