% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{save_dataset}
\alias{save_dataset}
\title{Save a dataset to JSON}
\usage{
save_dataset(dataset, path)
}
\arguments{
\item{dataset}{A \code{secureeval_dataset} object.}

\item{path}{File path to save to.}
}
\value{
\code{path} invisibly.
}
\description{
Save a dataset to JSON
}
