% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-reference.R
\name{reference_datasets}
\alias{reference_datasets}
\title{List available reference dataset names}
\usage{
reference_datasets()
}
\value{
Character vector.
}
\description{
List available reference dataset names
}
