% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securebench-package.R
\docType{package}
\name{securebench-package}
\alias{securebench}
\alias{securebench-package}
\title{securebench: Guardrail Benchmarking for Large Language Model Agents}
\description{
Benchmarking framework for guardrail accuracy in R Large Language Model (LLM) agent workflows. Evaluate guardrails against labeled datasets, compute precision, recall, and F1 metrics, generate confusion matrices, compare results across iterations, and export as 'vitals'-compatible scorers.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ian-flores.github.io/securebench/}
  \item \url{https://github.com/ian-flores/securebench}
  \item Report bugs at \url{https://github.com/ian-flores/securebench/issues}
}

}
\author{
\strong{Maintainer}: Ian Flores Siaca \email{iflores.siaca@hey.com} [copyright holder]

}
\keyword{internal}
